replaced old minesweeper with new version
